% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addKeysFromName.R
\name{addKeysFromName}
\alias{addKeysFromName}
\title{Dissociate the name of flow frames into multiple keys}
\usage{
addKeysFromName(flowObj, keys)
}
\arguments{
\item{flowObj}{The fcs object to be transformed. Both flowFrames and flowSets
are accepted.}

\item{keys}{A list of any number of characteristics that can be derived
from the sample names. For each entry, a gsub specification of where to find
the information in the file name should be added, such as id="..._|".}
}
\value{
The appended flowObj
}
\description{
Adds new keys to flow objects by separating the flowFrame 
identifiers/sample names into multiple entities.
}
