% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allHist.R
\name{allHist}
\alias{allHist}
\title{Plotting all variables as histograms}
\usage{
allHist(flowData, plotName = "all_histograms", nRows = 1e+05,
  saveResult = TRUE)
}
\arguments{
\item{flowData}{This is the full dataset, either a flowFrame or a flowSet, 
that should be plotted. If it has more rows than "nRows", a subsample (with 
equal contributions from each sample if a flowSet) will be plotted.}

\item{nRows}{The number of rows that will be used to construct the plot. 
The fewer, the faster, but the resolution also decreases, naturally. Default
is 100000.}

\item{saveResult}{Should the result be saved as a file?}
}
\value{
A plot with one histogram for each column in the flowData.
}
\description{
This function individually plots all histograms.
}
