% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendFFCols.R
\name{appendFFCols}
\alias{appendFFCols}
\title{This function is a slightly adjusted copy of the fr_append_cols
function in flowCore. It mainly allows for the presence of additional
pData variables}
\usage{
appendFFCols(focusFrame, newCols)
}
\description{
This function is a slightly adjusted copy of the fr_append_cols
function in flowCore. It mainly allows for the presence of additional
pData variables
}
