% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcTrans.R
\name{arcTrans}
\alias{arcTrans}
\title{Efficient inverse hyperbolic cosine transformation}
\usage{
arcTrans(flowObj, transNames = "default", exclColNmStr = c("ime"),
  transCoFacs = "default", unTrans = FALSE)
}
\arguments{
\item{flowObj}{The fcs object to be transformed. Both flowFrames and flowSets
are accepted.}

\item{transNames}{The variables that should be normalized. Default is all
that do not fall into the excl category below.}

\item{exclColNmStr}{A vector of strings that are not accepted in the
transNames.}

\item{transCoFacs}{This vector of values define the values for the
transformation during the normalization. In the "default" case, the function
defines the object as a CyTOF object if >5 percent of the values are 0, and 
applies the transformation value 8. Otherwise, the value 256 is applied.}

\item{unTrans}{If the reverse action should be taken, i.e. if an already 
transformed dataset should be un-transformed. NB! It is of great importance
that the same transformation factors are used!}
}
\value{
A flow object containing the transformed data, and with all metadata
left untouched.
}
\description{
This is a simple wrapper function for the base asinh function, that is useful
for flowFrames and flowSets. It also allows for reversing the transformation
with the argument reverse
}
