% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchNormCtrlSetup.R
\name{batchNormCtrlSetup}
\alias{batchNormCtrlSetup}
\title{Setup the file used for batch normalisation}
\usage{
batchNormCtrlSetup(ctrlFrame, normNames = colnames(ctrlFrame),
  exclColNmStr = c("ime"), volThresh = 0.05, lowVertThresh = 0.25,
  transCoFacs = "default", nonLevel1Var)
}
\arguments{
\item{ctrlFrame}{A control flowFrame that all the batches will be normalized
to.}

\item{normNames}{The variables that should be normalized. Default is all that
do not fall into one of the two excl categories below.}

\item{exclColNmStr}{A vector of strings that are not accepted in any
normNames.}

\item{volThresh}{Defining how small the smaller of the two peaks can be to be
considered a true peak. It is a fraction of the volume of the larger peak.
Default is 0.05, i.e. if the volume of the second peak is 5 percent or larger
than the volume of the first peak, it is considered a peak.}

\item{lowVertThresh}{Defining how low the turnpoint needs to be between the
peaks for these to be considered well-separated. 25 percent of the lower
peak is the default.}

\item{transCoFacs}{This vector of named values define the values for the
transformation during the normalization. This is only applied internally, so
transformation needs to be performed afterwards, preferrably with individual
values for each channel. In the "default" case, the function defines the file
as a CyTOF file, and applies the transformation value 8, if >5 percent of the
values are 0. Otherwise, the value 256 is applied. NB. The entries need to be
named in the same way as the normNames to secure that the right factor is
added to each variable.}

\item{nonLevel1Var}{If any of the markers are known to be unreliable,
these will be excluded from being used as filters for other markers.}
}
\description{
This function is used to generate the batch normalization values
that ideally are based on an identical control that is included
in each experiment.
}
