% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correcMatCreate.R
\name{correcMatCreate}
\alias{correcMatCreate}
\title{Generate a correction matrix for cytometry data analysis}
\usage{
correcMatCreate(flowObj, corrNames = "default")
}
\arguments{
\item{flowObj}{The dataset that should be corrected.}
}
\value{
A symmetrical matrix of zeros with the right row- and column names.
}
\description{
This function aids the correct unmix function, to create a symmetrical
correction matrix that should be used together with a flowframe to correct
the errors of unmixing.
}
