% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvSet2LongDf.R
\name{csvSet2LongDf}
\alias{csvSet2LongDf}
\title{Convert a set of csv files to one long dataframe with all identifiers as
separate columns.}
\usage{
csvSet2LongDf(path = ".", idInfo, nRows = "all")
}
\arguments{
\item{idInfo}{A list of any number of characteristics that can be derived
from the file names. For each entry, a gsub specification of where to find
the information in the file name should be added, such as id=""..._|..."".}

\item{filePath}{The path to the files to be imported.}
}
\value{
A long data frame with one column per PMT/APD (or fluorochrome,
depending on the state of the imported files), one for the acquisition date
(for fcs files) and one colum for each specified slot above. If no
gsub-pattern is provided, only a single column with the full file name will
be used to separate the observations from each file.
}
\description{
This function is mainly used for compatibility with DepecheR. It is analogous
to flowSet2Df, but takes files that have been pre-processed in e.g. FlowJo
instead of a flowSet.
}
