% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSet2LongDf.R
\name{flowSet2LongDf}
\alias{flowSet2LongDf}
\title{Convert a flowSet to one long dataframe with all identifiers as separate
#columns.}
\usage{
flowSet2LongDf(flowSetRaw, idInfo)
}
\arguments{
\item{idInfo}{A list of any number of characteristics that can be derived
from the file names. For each entry, a gsub specification of where to find
the information in the file name should be added, such as id=""..._|..."".}

\item{frames}{The flowSet to be converted. Can contain one single flowFrame.}
}
\value{
A long data frame with one column per PMT/APD (or fluorochrome,
depending on the state of the imported files), one for the acquisition date
(for fcs files) and one colum for each specified slot above. If no
gsub-pattern is provided, only a single column with the full file name will
be used to separate the observations from each file.
}
\description{
This function is mainly used for compatibility with DepecheR.
}
