% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullPanel.R
\docType{data}
\name{fullPanel}
\alias{fullPanel}
\title{Fully stained sample}
\format{An object of class \code{"data.frame"};}
\usage{
data(fullPanel)
}
\description{
This is a full PBMC sample stained with 12 fluorochromes detecting:
CD45RO (FITC), CD62L (APC), CD20 AF700, CD16 APC-Cy7, CD27 Qdot605,
CD45RA Qdot655, CD8 Qdot705, CD3 BrV785, CD38 PE, CD57 PE-CF594,
HLA-DR PE-Cy5.5, CD56 PE-Cy7. For this sample as for the compensation controls,
all 18 channels where left open when acquired on the BD Fortessa.
Date of  acquisition: 29.07.2015.
}
\keyword{datasets}
