% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating1D.R
\name{gating1D}
\alias{gating1D}
\title{importFrom stats density}
\usage{
gating1D(dataSet, variable, outputDataset, gateName = variable,
  color = "red", saveResult = TRUE)
}
\description{
importFrom stats density
}
