% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneDZeroTrim.R
\name{oneDZeroTrim}
\alias{oneDZeroTrim}
\title{importFrom flowCore exprs
importFrom BiocGenerics nrow ncol colnames}
\usage{
oneDZeroTrim(focusFrame, trimFrac, nRowOut, returnFlowFrame = TRUE,
  returnPositions = FALSE)
}
\description{
importFrom flowCore exprs
importFrom BiocGenerics nrow ncol colnames
}
