% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgePlot.R
\name{ridgePlot}
\alias{ridgePlot}
\title{Density plots showing gates for all files in a flowSet}
\usage{
ridgePlot(focusSet, plotVar, filtVar, sampleKey = "GUID",
  plotName = plotVar, saveResult = TRUE)
}
\arguments{
\item{focusSet}{The flow set in focus.}

\item{plotVar}{The variable that should be plotted.}

\item{filtVar}{The variable where the filter is present.}

\item{sampleKey}{The name on the keyword from which the names of the samples
can be extracted.}

\item{plotName}{If a name different from plotVar should be used, it can be
added here.}

\item{saveResult}{Should the result be saved as a file?}
}
\description{
This function is completely analogous to the the densityplot function in
flowViz, but it takes the infor straight from the flowFrame, without the need
for any filter functions or objects, and the interface is simpler.
}
