% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compCtrls.R
\docType{data}
\name{compCtrls}
\alias{compCtrls}
\title{Compensation controls}
\format{An object of class \code{"flowSet"}}
\usage{
data(compCtrls)
}
\description{
This is a flowSet with 14 compensation controls:
12 single-stained bead populations, 1 unstained bead and 1 PBMC control,
experimentally treated the same way as the "fullPanel" file. 2000 events
from each file are included. All 18 channels where left open when acquired
on the BD Fortessa. Date of  acquisition: 29.07.2015.
}
\keyword{datasets}
