% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFlowSetDf.R
\name{readFlowSetDf}
\alias{readFlowSetDf}
\title{Import a flow set and convert it to one long dataframe}
\usage{
readFlowSetDf(path = ".", id, group, stim)
}
\arguments{
\item{path}{The path to the folder where the .fcs files are stored.}

\item{id}{If a column specifying which of the original donors that each observation comes from should be included, the gsub-pattern should be specified here.}

\item{group}{See "ids"}

\item{stim}{See "ids"}
}
\value{
A long data frame with one column per PMT/APD, one for the acquisition date and one colum for each specified slot above. If the same donor id occurs for two different stimuli, another column creating unique ids is also made. If no gsub-pattern is provided, the full file name will be used to separate the observations.
}
\description{
This function is useful to manipulate data within R, as, in most instances, the majority of the meta information stored in the fcs file is not of general use for day.to.day analyses.
}
