% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specMatCalc.R
\name{specMatCalc}
\alias{specMatCalc}
\title{Calculating the matrix used for spectral unmixing}
\usage{
specMatCalc(compControls, compScheme)
}
\arguments{
\item{compControls}{A dataframe with two components (and many columns):
\describe{
           \item{Name}{A column that associates each observation with a certain sample}
           \item{Channels}{All the channels detecting fluosrescent information in the file.}
           \item{Negative}{The file number in the compControls list/FlowSet negative for the fluorochrome/marker in question. For autofluorescense subtraction, the value here is set to "NA".}
         }}

\item{compScheme}{A dataframe with three columns:
\describe{
           \item{Fluorochrome}{The name of the fluorochrome/marker present in the Name column of the compControls file}
           \item{Positive}{The file number in the compControls list/FlowSet positive for the fluorochrome/marker in question.}
           \item{Negative}{The file number in the compControls list/FlowSet negative for the fluorochrome/marker in question. For autofluorescense subtraction, the value here is set to "NA".}
         }}
}
\value{
A data frame with each row representing a fluorochrome and each column representing a detector
}
\description{
This algoritm takes the (pre-gated) single-stained controls and negative controls, including an autofluorescence control and estimates the unmixing for all fluorescent variables.
}
