% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specUnmix.R
\name{specUnmix}
\alias{specUnmix}
\title{Plotting all variables against a single variable}
\usage{
specUnmix(rawData, specMat)
}
\arguments{
\item{rawData}{This can be either be a data frame, a FlowFrame or a FlowSet.}

\item{specMat}{Here, the matrix generated by the secMatCalc function is added, that is used for the}
}
\value{
The unmixed data. It will be returned in the format it was imported as.
}
\description{
This function is useful both when setting appropriate gates and when the adjustments of the compensation are done
}
